% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClearSkyQuantileRegression.R
\name{aggregate2yrmon}
\alias{aggregate2yrmon}
\title{Aggregate sub-daily data first to a monthly mean diurnal cylce and then to monthly average}
\usage{
aggregate2yrmon(dth = NULL, Date = NULL, Time = NULL,
  IncomingShortwave = NULL, ..., nminmon = 26)
}
\arguments{
\item{dth}{a data.table with column names Date, Time; if not supplied it will be constructed from vectors}

\item{Date}{a vector of Dates in the format of data.table::as.IDate}

\item{Time}{a vector of Times in the format of data.table::as.ITime}

\item{IncomingShortwave}{a vector of subdaily (hourly or half-hourly) observations of solar radiation}

\item{...}{further vectors which shall be aggregated}

\item{nminmon}{numeric, sets the required number of existing days for averaging}
}
\value{
data.table with aggregated data per year and month
}
\description{
The aggregation tries to avoid sampling biases (e.g. missing data at night time)
with 26 days required and then average the monthly mean diurnal cycle
}
\details{
Provide either a data table dth or vectors of Date, Time, IncomingShortwave
}
\author{
Maik Renner, mrenner [at] bgc-jena.mpg.de
}
