% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClearSkyQuantileRegression.R
\name{calc_ClearSky_QuantileRegression}
\alias{calc_ClearSky_QuantileRegression}
\title{Calculate fractional absorption of solar radiation by Quantile regression
of the observed incoming shortwave radiation against potential radiation}
\usage{
calc_ClearSky_QuantileRegression(IncomingShortwave,
  IncomingShortwavePotential, tau = 0.85)
}
\arguments{
\item{IncomingShortwave}{Vector of observed incoming total Shortwave radiation in units W m-2, also refferred to as global radiation or the sum of diffuse and direct radiation if available}

\item{IncomingShortwavePotential}{Vector pf potential surface shortwave radiation at the surface without clouds.}

\item{tau}{Numeric constant, quatile for which the regression is being performed.
Defaults to 0.85 based on many BSRN stations, but this may vary (cloud type, frequency and aerosols).}
}
\value{
A data.table with columns tau, statistic and value.
Key statistics are R1 (measure of goodness of fit), slope1 (linear slope assumed to represent the fractional absorption of solar radiation under cloud-free conditions),
slope_sd (the standard deviation of the slope estimate), intercept (which should be close to 0, otherwise there is a problem in the correct timing of the observed and potential radiation), and
n the sample size
}
\description{
The slope at a given quantile (here 0.85) is assumed to represent the fractional absorption of solar radiation under cloud-free conditions.
This can be used to estimate the clear sky solar radiation of a period of one month when global radiation records are available.
The observations should not be biased, since this directly effects the estimate.
Persistent cloud-cover will reduce the estimate, but often results in poorer goodness of fit.
The estimate is not useful for nighttime conditions.
Details are described in Renner et al., 2019 ESS
}
\references{
Renner, M., M. Wild, M. Schwarz, and A. Kleidon.
  "Estimating Shortwave Clear-Sky Fluxes from Hourly Global
    Radiation Records by Quantile Regression."
      Earth and Space Science, 2019.
      \url{https://doi.org/10.1029/2019EA000686}
}
\seealso{
\code{[calc_PotRadiation_CosineResponsePower()]} for calculating  potential surface shortwave radiation as function of date, time and position
}
\author{
Maik Renner, mrenner [at] bgc-jena.mpg.de
}
