% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClearSkyQuantileRegression.R
\name{calc_ClearSky_QuantileRegression_MonthlyTimeWindow}
\alias{calc_ClearSky_QuantileRegression_MonthlyTimeWindow}
\title{Main function to estimate the clear-sky solar radiation based on (half-)hourly global radiation data for monthly periods.}
\usage{
calc_ClearSky_QuantileRegression_MonthlyTimeWindow(Date, Time,
  IncomingShortwave, IncomingShortwavePotential = NULL, tau = 0.85,
  lat = NULL, lon = NULL, hourshift = 0.25, timeZone = 0,
  isCorrectSolartime = TRUE, cosineResponsePower = 1.2, pdev = 0.25,
  mc.cores = 1, briefoutput = TRUE)
}
\arguments{
\item{Date}{a vector of Dates in the format of data.table::as.IDate}

\item{Time}{a vector of Times in the format of data.table::as.ITime}

\item{IncomingShortwave}{a vector of subdaily (hourly or half-hourly) observations of solar radiation}

\item{IncomingShortwavePotential}{a vector of potential solar radiation (if not supplied it will be computed)}

\item{tau}{numeric the quatile for which the regression is being performed
Defaults to 0.85 based on many BSRN stations, but this may vary (cloud type, frequency and aerosols)}

\item{lat}{Latitude of the site in degrees}

\item{lon}{Longitude of the site in degrees}

\item{hourshift}{numeric A shift in hours to align the potential solar radiation with the observed radiation. Here 0.25 is used as default for half-hourly aggregated data
(a shift will induce hyseresis and an intercept term in quantile regression)}

\item{timeZone}{numeric, if local time is used then 0 otherwise the time must be adapted}

\item{cosineResponsePower}{numeric defaults to 1.2, may be lower at high latitude sites, see Long and Ackerman (2000) JGR}

\item{pdev}{numeric allowable fractional deviation of the monthly estimate from the site average estimate (here 0.25)}

\item{mc.cores}{Numeric setting the number of parallel child processes to calculate the regressions for each month, \code{@seealso[detectCores()]}}

\item{briefoutput}{logical defaults to TRUE, otherwise all statistics are reported}
}
\value{
a data.table with average fluxes for year and months, the fractional transmission estimate (ftau), its corresponding Window length and quantile and the monthly mean ClearSky incoming solar radiation
}
\description{
This function will estimate the fractional solar transmission per year and month and for moving averages with window sizes of 3, 5, and 7 months.
  Based on the deviation of the monthly estimates from the site average and the goodness of fit the appropriate window is selected.
  For this window the resulting fractional solar transmission is used to calculate the ClearSky Shortwave radiation.
  The methodology is described in Renner et al., (2019) ESS.
  The function may take some time since different window length are computed.
}
\details{
This function may also compute the potential solar radiation when this is not given, this requires lat, long, time with timezone and possible shift.

To speed up the calculations the parallel package is used. This will not work under Windows, unless mc.cores = 1.
}
\examples{
data(LIN2003)
# apply regression per month and with different windows
# Check available cores to parallize the task with setting mc.cores larger than 1
detectCores()
(rqmw = LIN2003[ , calc_ClearSky_QuantileRegression_MonthlyTimeWindow(Date,Time,IncomingShortwave, tau = 0.85, lat = 52.21, lon = 14.122, hourshift = 0.5,timeZone = 0, mc.cores = 1)])

plot(IncomingShortwavePotential ~ month, data = rqmw, type = "l", col = 4, ylab = "Shortwave Radiation (W/m2)", ylim = c(0,500))
lines(IncomingShortwaveClearSky ~ month, data = rqmw, type = "l", col =2)
lines(IncomingShortwave ~ month, data = rqmw, type = "b")
legend("topright", c("Potential", "Clear Sky flux", "Observed at surface"), col = c(4,2,1), lty = 1)


}
\references{
Renner, M., M. Wild, M. Schwarz, and A. Kleidon.
  "Estimating Shortwave Clear-Sky Fluxes from Hourly Global
    Radiation Records by Quantile Regression."
      Earth and Space Science, 2019.
      \url{https://doi.org/10.1029/2019EA000686}
}
\author{
Maik Renner, mrenner [at] bgc-jena.mpg.de
}
