% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PotentialRadiation.R
\name{calc_PotRadiation_CosineResponsePower}
\alias{calc_PotRadiation_CosineResponsePower}
\title{Model potential solar radiation at the surface with power law to remove non-linearities with Incoming solar}
\usage{
calc_PotRadiation_CosineResponsePower(doy, hour, latDeg, longDeg, timeZone,
  isCorrectSolartime = TRUE, cosineResponsePower = 1.2)
}
\arguments{
\item{doy}{value or vector of day of year}

\item{hour}{value or vector of hour of day also with decimal fraction}

\item{latDeg}{latitude in degrees}

\item{timeZone}{numeric, if local time is used then 0 otherwise the time must be adapted}

\item{cosineResponsePower}{numeric defaults to 1.2, may be lower at high latitude sites, see Long and Ackerman 2000}

\item{latDeg}{longitude in degrees}
}
\value{
vector of potential solar radiation at the surface
}
\description{
Attenuation of the solar beam in a curved, refractive atmosphere modelled with a simple power
law with an exponent to the cosine of the solar zenit angle
The function calles other functions to calculate the solar zenith angle, the extra terrestrial
radiation copied from  the R packages
REddyProc::fCalcPotRadiation and solartime.
Since their arguments where changing I copied part of these packages.
}
\examples{
data(LIN2006)
# calculate potential surface solar radiation
LIN2006[ , IncomingShortwavePotential := calc_PotRadiation_CosineResponsePower(doy = yday(Date),hour = Time/3600 + 0.5, latDeg = 52.21, longDeg = 14.122, timeZone = 0) ]
LIN2006
#plot(IncomingShortwave ~ IncomingShortwavePotential, data = LIN2006[month(Date) == 6 &  mday(Date) == 7, ], type = "l")

plot(IncomingShortwave ~ IncomingShortwavePotential, data = LIN2006[month(Date) == 6, ], type = "p")
}
\references{
Long and Ackerman 2000 JGR
Renner et al., 2019 ESS
}
\author{
Maik Renner, mrenner [at] bgc-jena.mpg.de
}
