% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_table_quantile_regression_utils.R
\name{mrq.output.statlong}
\alias{mrq.output.statlong}
\title{Reorganize the output of a quantile regression fit into a long table format: statistic, value}
\usage{
mrq.output.statlong(fit, se = "iid")
}
\arguments{
\item{fit}{object returned from a linear model quantreg::rq() are supported}
}
\value{
a data.table with columns statistic and value
}
\description{
simple function which takes a lm model output to create a named vector intended for use with data.table and group by
}
\examples{
# output of regression model is melted to two columns
library(quantreg)
data(engel)
fit = rq(foodexp~income,tau=0.9,data=engel)
mrq.output.statlong(fit)
}
\author{
Maik Renner, mrenner [at] bgc-jena.mpg.de
}
